persistance now works both ways: vcell to database and back
