+++ vcell-core/src/test/java/cbit/vcell/biomodel/SpringSaLaDGoodReactionsTest.java
@@ -263,70 +263,6 @@ public class SpringSaLaDGoodReactionsTest {
         assertTrue((solver instanceof LangevinSolver) ? true : false, "expecting instanceof Langevin solver");
 	}
 
-	/* -------------------------------------------------------------------------------------------------
-	 * This test will exercise creation and parsing of a species context spec entry in the SQL table
-	 *
-	 */
-	@Test
-	public void test_springsalad_application() throws IOException, XmlParseException, PropertyVetoException, ExpressionException, GeometryException,
-			ImageException, IllegalMappingException, MappingException, SolverException {
-
-		BioModel bioModel = getBioModelFromResource("Spring_application.vcml");
-		SimulationContext simContext = bioModel.getSimulationContext(0);
-
-		// WARNING!! Debug configuration for this JUnit test required System property "vcell.installDir"
-		// ex: -Dvcell.installDir=C:\dan\jprojects\git\vcell
-		MathDescription mathDescription = simContext.getMathDescription();
-        assertTrue((mathDescription.getMathType() != null && mathDescription.getMathType() == SpringSaLaD) ? true : false, "expecting SpringSaLaD math type");
-
-		SpeciesContextSpec[] speciesContextSpecs = simContext.getReactionContext().getSpeciesContextSpecs();
-		SpeciesContextSpec scs = speciesContextSpecs[0];		// we test roundtrip for just one SpeciesContextSpec
-		String internalLinkSetSQL = scs.getInternalLinksSQL();
-		String siteAttributesMapSQL = scs.getSiteAttributesSQL();
-		Set<MolecularInternalLinkSpec> internalLinkSet = scs.readInternalLinksSQL(internalLinkSetSQL);
-		Map<MolecularComponentPattern, SiteAttributesSpec> siteAttributesMap = scs.readSiteAttributesSQL(siteAttributesMapSQL);
-
-		double bondLength = 1;
-		Map<ReactionRuleSpec.Subtype, Integer> subTypeMap = new LinkedHashMap<>();
-		ReactionRuleSpec[] rrSpecs = simContext.getReactionContext().getReactionRuleSpecs();
-		for(ReactionRuleSpec rrs : rrSpecs) {
-			Map<String, Object> analysisResults = new LinkedHashMap<>();
-			rrs.analizeReaction(analysisResults);
-			ReactionRuleSpec.Subtype st = rrs.getSubtype(analysisResults);
-			if(subTypeMap.containsKey(st)) {
-				int count = subTypeMap.get(st);
-				count++;
-				subTypeMap.put(st, count);
-			} else {
-				subTypeMap.put(st, 1);
-			}
-			if(ReactionRuleSpec.Subtype.BINDING == st) {
-				bondLength = rrs.getFieldBondLength();
-			}
-		}
-        assertTrue(subTypeMap.size() == 5 ? true : false, "Number of compatible subtypes must be 5");
-        assertTrue(subTypeMap.containsKey(INCOMPATIBLE) ? false : true, "No incompatible subtype may exist");
-        assertTrue(bondLength == 3 ? true : false, "BondLength must be 3");
-
-		// verify roundtrip for internalLinkSet (through sampling)
-        assertTrue(internalLinkSet.size() == scs.getInternalLinkSet().size() ? true : false, "internalLinkSet size different after roundtrip");
-		MolecularInternalLinkSpec milsThis = internalLinkSet.iterator().next();
-		boolean found = false;
-		for(MolecularInternalLinkSpec milsThat : scs.getInternalLinkSet()) {
-			if(milsThis.compareEqual(milsThat)) {
-				found = true;
-				break;
-			}
-		}
-        assertTrue(found ? true : false, "MolecularInternalLinkSpec element not found after roundtrip");
-
-		// verify roundtrip for siteAttributesMap (through sampling)
-        assertTrue(siteAttributesMap.size() == scs.getSiteAttributesMap().size() ? true : false, "siteAttributesMap size different after roundtrip");
-		MolecularComponentPattern mcpThis = siteAttributesMap.keySet().iterator().next();
-		SiteAttributesSpec sasThis = siteAttributesMap.get(mcpThis);
-		SiteAttributesSpec sasThat = scs.getSiteAttributesMap().get(mcpThis);
-        assertTrue(sasThis.compareEqual(sasThat) ? true : false, "SiteAttributesSpec element not found in siteAttributesMap after roundtrip");
-	}
 
 	/* -------------------------------------------------------------------------------------------------
 	 * This test will exercise loading a springsalad mathmodel
